Move CacheException into Exception directory, keeping BC with class_alias shim
