+++ namespace-bc-aliases.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias(\Roave\DoctrineSimpleCache\Exception\CacheException::class, \Roave\DoctrineSimpleCache\CacheException::class);
